Add an opensource Qualcomm Shortcut FE driver.
The driver consists of:
- sfe (core modules): shortcut-fe (ipv4) and shortcut-fe-ipv6 -
modules that actually do the offloading
- shortcut-fe-cm - simple shortcut-fe connection manager - a
module that creates offloading rules and passes to sfe
- fast-classifier - advanced connection manager - a module that
creates offloading rules and passes to sfe. Comparing to sfe-cm
has additional checks, connection statistics and creates an
offloading rule only when a connection hits 128 packets flowed
(can be adjusted through /sys/fast_classifier/offload_at_pkts)
- userspace-fast-classifier - a fast-classifier user-space tool
to manually add connections one want to offload
Both sfe-cm and fast-classifier get information about connections
through netfilter. And after some checks it creates an offloading
rule that is passed to sfe. After that sfe marks a connection as
fast-forwarded and routes the flow.

Kernel parses this mark and bypasses its network stack taps for
a specific connection in favor of sfe.

There are 3 patches added with this commit:
950 - adds a hook into network stack, sets no tcp window check
to 0 (according to sfe readme).
951 - because fast-classifier also supports bridge offloading
(echo 1 > /sys/fast_classifier/skip_to_bridge_ingress) and
offloaded connections do not naturally update bridge statistics,
fast-classifier will do that.
952 - allow multiple listeners for conntrack events, it is
requied for netlink and sfe coexistance (for example)

All these patches are filled with ifdefs so the sfe code path
is engaged only when these modules are compiled. It allows to
keep kernel code untouched in case of no sfe compiled in.

Because sfe-cm and fast-classifier are competing modules, it's
recommended to keep loaded only one of them. If both are loaded
then in a race condition sfe-cm always comes first.

The driver is a merger of
https://source.codeaurora.org/quic/qsdk/oss/lklm/shortcut-fe/log/?h=release/endive_cc
with
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/chromeos-3.18/drivers/net/ethernet/qualcomm/sfe/

Also added missing function to update udp statistics by
fast-classifier and some more small fixes.

Patches are taken from QSDK and that chromeos-3.18 repo and a bit
adjusted.

Signed-off-by: Pavel Kubelun <be.dissent@gmail.com>
Signed-off-by: Sebastian Careba <sebi@rnp.rosilva.ro>

Tested by: Sebastian Careba <sebi@rnp.rosilva.ro>